Add basic url for answer. Also, seperate test files for further implementation.
