+++ api/urls.py
@@ -11,6 +11,8 @@ urlpatterns = [
     url(r'posts/index$', posts_views.index, name='index_posts'),
     url(r'posts/question$', posts_views.question, name='question_posts'),
     url(r'posts/question/(?P<id>[0-9]+)$', posts_views.question, name='question_post_by_id'),
+    url(r'posts/answer$', posts_views.answer, name='question_answers'),
+    url(r'posts/answer/(?P<id>[0-9]+)$', posts_views.answer, name='question_answers_by_id'),
     url(r'posts/information$', posts_views.information, name='information_posts'),
     url(r'posts/information/(?P<id>[0-9]+)$', posts_views.information, name='information_post_by_id'),
     url(r'list/questions$', list_views.questions, name='question_list'),
